refactor: use shared empty state for stock list
