+++ src/main/java/edu/ntnu/idatt2003/millions/view/component/card/PortfolioValueCard.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idatt2003.millions.view.component;
+package edu.ntnu.idatt2003.millions.view.component.card;
 
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.service.PortfolioService;
