+++ src/main/java/edu/ntnu/idatt2003/millions/view/exchange/overview/card/GainersCard.java
@@ -2,7 +2,7 @@ package edu.ntnu.idatt2003.millions.view.exchange.overview.card;
 
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.view.component.StyledText;
-import edu.ntnu.idatt2003.millions.view.component.WidgetCard;
+import edu.ntnu.idatt2003.millions.view.component.card.WidgetCard;
 
 /**
  * Widget card displaying the number of stocks that rose in price this week.
