+++ src/main/java/edu/ntnu/idatt2003/millions/view/exchange/stocks/StocksView.java
@@ -3,9 +3,9 @@ package edu.ntnu.idatt2003.millions.view.exchange.stocks;
 import edu.ntnu.idatt2003.millions.controller.PortfolioController;
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.util.LanguageManager;
-import edu.ntnu.idatt2003.millions.view.component.AvailableFundsCard;
+import edu.ntnu.idatt2003.millions.view.component.card.AvailableFundsCard;
 import edu.ntnu.idatt2003.millions.view.component.Pagination;
-import edu.ntnu.idatt2003.millions.view.component.PortfolioValueCard;
+import edu.ntnu.idatt2003.millions.view.component.card.PortfolioValueCard;
 import edu.ntnu.idatt2003.millions.view.component.SearchBar;
 import edu.ntnu.idatt2003.millions.view.component.StyledText;
 import edu.ntnu.idatt2003.millions.view.exchange.stocks.card.StocksInvestedCard;
