+++ src/main/java/edu/ntnu/idatt2003/millions/view/exchange/stocks/card/StocksInvestedCard.java
@@ -4,7 +4,7 @@ import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.service.PortfolioService;
 import edu.ntnu.idatt2003.millions.util.CurrencyFormatter;
 import edu.ntnu.idatt2003.millions.util.LanguageManager;
-import edu.ntnu.idatt2003.millions.view.component.SimpleWidgetCard;
+import edu.ntnu.idatt2003.millions.view.component.card.SimpleWidgetCard;
 
 import java.text.MessageFormat;
